+++ CHANGELOG.md
@@ -3,12 +3,10 @@
 ## [Unreleased]
 ### Added
 * Add button to Bugzilla on public and private comments
+* `DEFER` is now a possible affect resolution, but can only be selected when impact is `LOW` (`OSIDB-3285`, `OSIDB-3286`, `OSIDB-3288`)
 
 ### Fixed
 * Allow saving flaws with historical affects (`OSIDB-3262`)
-
-### Changed
-* `DEFER` has replaced `WONTFIX` as an affect resolution, and cannot be selected when impact is not `LOW` (`OSIDB-3285`, `OSIDB-3286`, `OSIDB-3288`)
 
 ## [2024.7.2]
 ### Fixed